rewriting into hooks
